MIgrate expense pages
