+++ packages/cypress/cypress/e2e/Expenses/edit-expense.spec.ts
@@ -150,7 +150,9 @@ describe('Edit expense', () => {
         const newAmount =
           targetTransactionBefore.amount + amountToChangeTransaction;
 
-        cy.visit('http://localhost:3000/statistics/expenses/2021-02/1');
+        cy.visit(
+          'http://localhost:3000/statistics/expenses?date=2021-02&page=1'
+        );
 
         cy.getById(targetTransactionBefore._id).click();
