+++ packages/frontend/containers/expenses/add-expense.container.tsx
@@ -1,26 +1,35 @@
 import { useRouter } from 'next/router';
 import { useState } from 'react';
 
-import { ExpenseForm } from './ExpenseForm';
-
 import {
   CreateExpenseDto,
-  TransactionTypeEnum,
   useExpensesCreateMutation,
+  useTransactionTemplatesFindOneQuery,
 } from '$api/generated/financerApi';
-import { TransactionTemplateSwitcher } from '$blocks/transaction-template-switcher/transaction-template-switcher';
-import { Loader } from '$elements/loader/loader';
-import { LoaderFullScreen } from '$elements/loader/loader.fullscreen';
+import { DataHandler } from '$blocks/data-handler/data-handler';
 import { useUserDefaultExpenseAccount } from '$hooks/profile/user-preference/useUserDefaultExpenseAccount';
-import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
+import { AddExpense } from '$pages/expenses/add-expense';
 import { parseErrorMessagesToArray } from '$utils/apiHelper';
 
-export const AddExpense = (): JSX.Element => {
+interface AddExpenseContainerProps {
+  templateId?: string;
+}
+
+export const AddExpenseContainer = ({
+  templateId,
+}: AddExpenseContainerProps) => {
   const { push } = useRouter();
   const [errors, setErrors] = useState<string[]>([]);
   const [addExpense, { isLoading: isCreating }] = useExpensesCreateMutation();
   const { data: defaultExpenseAccount, isLoading: isLoadingDefaultAccount } =
-    useUserDefaultExpenseAccount();
+    useUserDefaultExpenseAccount({ skip: !!templateId });
+
+  const templateData = useTransactionTemplatesFindOneQuery(
+    { id: templateId as string },
+    { skip: !templateId }
+  );
+
+  const { data: transactionTemplate } = templateData;
 
   const handleSubmit = async (newExpenseData: CreateExpenseDto) => {
     try {
@@ -45,23 +54,15 @@ export const AddExpense = (): JSX.Element => {
 
   return (
     <>
-      {isCreating && <LoaderFullScreen />}
-      <UpdatePageInfo
-        title="Add expense"
-        headerAction={
-          <TransactionTemplateSwitcher
-            templateType={TransactionTypeEnum.Expense}
-          />
-        }
-      />
-      {isLoading ? (
-        <Loader />
-      ) : (
-        <ExpenseForm
-          onSubmit={handleSubmit}
+      <DataHandler skipNotFound {...templateData} />
+      {(!templateId || transactionTemplate) && (
+        <AddExpense
+          defaultExpenseAccount={defaultExpenseAccount}
+          expenseTemplate={transactionTemplate}
+          isLoading={isLoading}
+          isCreating={isCreating}
           errors={errors}
-          submitLabel="Submit"
-          initialValues={{ fromAccount: defaultExpenseAccount }}
+          onSubmit={handleSubmit}
         />
       )}
     </>
