+++ packages/frontend/pages/statistics/expenses/[expenseId]/edit.tsx
@@ -1,0 +1,13 @@
+import { useRouter } from 'next/router';
+
+import { EditExpenseContainer } from '$container/expenses/edit-expense.container';
+
+const EditExpensePage = () => {
+  const {
+    query: { expenseId },
+  } = useRouter();
+
+  return <EditExpenseContainer id={expenseId as string} />;
+};
+
+export default EditExpensePage;
