+++ packages/frontend/pages/statistics/expenses/add/index.tsx
@@ -1,0 +1,7 @@
+import { AddExpenseContainer } from '$container/expenses/add-expense.container';
+
+const AddExpensePage = () => {
+  return <AddExpenseContainer />;
+};
+
+export default AddExpensePage;
