+++ packages/frontend/pages/statistics/expenses/index.tsx
@@ -1,0 +1,18 @@
+import { useRouter } from 'next/router';
+
+import { ExpenseListingContainer } from '$container/expenses/expense-listing.container';
+
+const ExpensesPage = () => {
+  const {
+    query: { date, page },
+  } = useRouter();
+
+  return (
+    <ExpenseListingContainer
+      date={date as string}
+      page={parseInt(page as string)}
+    />
+  );
+};
+
+export default ExpensesPage;
