+++ packages/frontend/src/pages/incomes/__test__/groupIncomesByMonth.test.ts
